test/integration/server/requirements.txt
grpcio==1.67.0
grpcio-tools==1.67.0
iniconfig==2.0.0
packaging==24.1
pluggy==1.5.0
protobuf==5.27.2
pytest==8.3.3
